Create authentication with github and memory CRUD routes.

+++ server/src/routes/memories.ts
@@ -3,8 +3,15 @@ import { prisma } from '../lib/prisma'
 import { z } from 'zod'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get('/memories', async () => {
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify()
+  })
+
+  app.get('/memories', async (request) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: 'asc',
       },
@@ -18,17 +25,22 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
   })
 
-  app.get('/memories/:id', async (request) => {
+  app.get('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
     const { id } = paramsSchema.parse(request.params)
 
-    const memory = prisma.memory.findUniqueOrThrow({
+    const memory = await prisma.memory.findUniqueOrThrow({
       where: {
         id,
       },
     })
+
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send
+    }
+
     return memory
   })
   app.post('/memories', async (request) => {
@@ -45,12 +57,12 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: '7608ca58-9cc5-4ce1-8f90-0c0a167590f1',
+        userId: request.user.sub,
       },
     })
     return memory
   })
-  app.put('/memories/:id', async (request) => {
+  app.put('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -64,8 +76,16 @@ export async function memoriesRoutes(app: FastifyInstance) {
     const { id } = paramsSchema.parse(request.params)
 
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
+    let memory = await prisma.memory.findUnique({
+      where: {
+        id,
+      },
+    })
 
-    const memory = await prisma.memory.update({
+    if (memory?.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -77,11 +97,21 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
     return memory
   })
-  app.delete('/memories/:id', async (request) => {
+  app.delete('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
     const { id } = paramsSchema.parse(request.params)
+
+    const memory = await prisma.memory.findUnique({
+      where: {
+        id,
+      },
+    })
+
+    if (memory?.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
 
     await prisma.memory.delete({
       where: {
